LayoutTests/fast/frames/xss-auditor-handles-file-urls-expected.txt

CONSOLE MESSAGE: line 3: The XSS Auditor refused to execute a script in 'html;base64,PHNjcmlwdD5hbGVydCgxKTwvc2NyaXB0Pg=='%3E' because its source code was found within the request. The auditor was enabled because the server did not send an 'X-XSS-Protection' header.
